Checkpointing before trying to rewrite some graph stuff
I changed `Source`, it now carries a table name. To make this work correctly I
will now try to change node types to `Operator` immediately so I can just add
the table name as well. This might fail, if earlier transformations rely on the
node type being `QualifiedBinding`, in which case we need to encode the table as
a `QualifiedBinding`, which is ugly.