Added table references, fixed some bugs
Table references let us load udfs without reqiring arguments
Added typing for if and select
Fixed that the `if` entry operator is `ifFun`, not `ifThenElse` in `inlineCtrl`
`sequentializeScalars` no longer requires the parent (_, One) semantic. This is
a trial and may also cause a bug, in future probably this should use typing.
Added `Union` removal in `removeSuperfluousOperators`
Added table refs to `retype3`
Tables can now be referenced by name or index.
Added a bit of `or` handling, but only in cases where it de-morgans back to
`and`
Fixed a bug where `inlineAlgos` would `take` an `MVar` instead of just
`read`'ing, thus loosing the namespace if it was needed twice